attached_assets/pyproject.toml
[project]
name = "emeralds-killfeed"
version = "0.1.0"
description = "Emerald's Killfeed Discord bot for Deadside PvP Engine"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "apscheduler>=3.11.0",
    "asyncssh>=2.21.0",
    "discord-py>=2.5.2",
    "motor>=3.7.1",
    "paramiko>=3.5.1",
    "py-cord==2.6.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "setuptools>=80.8.0",
]
